refactor: Move typewriter to a mixin
